Add sync option to sync data over multiple tabs.
Added an option to sync data from one tab to another using the `storage` event. This event gets triggered every time the localStorage tab gets updated on blurred tabs and can be used to notify another tab that data needs to be updated. We can then use this to retrieve the new value from the localStorage and update the values in the store.